Refactored InheritanceStructure constructor
