+++ graphql_compiler/schema_generation/orientdb/schema_graph_builder.py
@@ -8,7 +8,7 @@ import six
 from ..exceptions import IllegalSchemaStateError
 from ..schema_graph import (
     EdgeType, IndexDefinition, InheritanceStructure, NonGraphElement, PropertyDescriptor,
-    SchemaGraph, VertexType, get_subclass_sets_from_superclass_sets, link_schema_elements
+    SchemaGraph, VertexType, link_schema_elements
 )
 from .schema_properties import (
     COLLECTION_PROPERTY_TYPES, EDGE_DESTINATION_PROPERTY_NAME, EDGE_END_NAMES,
@@ -16,7 +16,6 @@ from .schema_properties import (
     ORIENTDB_BASE_VERTEX_CLASS_NAME, PROPERTY_TYPE_LINK_ID, UNIQUE_INDEX_TYPES,
     parse_default_property_value, try_get_graphql_scalar_type
 )
-from .utils import toposort_classes, toposort_classes
 
 
 def get_orientdb_schema_graph(schema_data, index_data):
@@ -93,10 +92,7 @@ def get_orientdb_schema_graph(schema_data, index_data):
         for class_name, class_definition in six.iteritems(class_name_to_definition)
     }
 
-    toposorted_class_to_immediate_superclasses = toposort_classes(class_to_immediate_superclasses)
-
-    inheritance_structure = _get_inheritance_structure_from_schema_data(
-        toposorted_class_to_immediate_superclasses)
+    inheritance_structure = InheritanceStructure(class_to_immediate_superclasses)
 
     non_graph_elements = _get_non_graph_elements(class_name_to_definition, inheritance_structure)
     inner_collection_objs = _get_graphql_representation_of_non_graph_elements(
@@ -120,32 +116,6 @@ def get_orientdb_schema_graph(schema_data, index_data):
     else:
         all_indexes = _get_indexes(index_data, elements)
     return SchemaGraph(elements, inheritance_structure, all_indexes)
-
-
-def _get_inheritance_structure_from_schema_data(toposorted_class_to_immediate_superclasses):
-    """Return the superclass sets from the OrientDB schema data."""
-    # For each class name, construct its superclass set:
-    # itself + the set of class names from which it inherits.
-    superclass_sets = dict()
-    for class_name, immediate_superclass_names in six.iteritems(
-            toposorted_class_to_immediate_superclasses):
-        superclass_set = set(immediate_superclass_names)
-        superclass_set.add(class_name)
-
-        # Since the input data must be in topological order, the superclasses of
-        # the current class should have already been processed.
-        # A KeyError on the following line would mean that the input
-        # was not topologically sorted.
-        superclass_set.update(chain.from_iterable(
-            superclass_sets[superclass_name]
-            for superclass_name in immediate_superclass_names
-        ))
-
-        # Freeze the superclass set so it can't ever be modified again.
-        superclass_sets[class_name] = frozenset(superclass_set)
-    subclass_sets = get_subclass_sets_from_superclass_sets(superclass_sets)
-    return InheritanceStructure(superclass_sets, subclass_sets)
-
 
 def get_superclasses_from_class_definition(class_definition):
     """Extract a set of all superclass names from a class definition dict."""
