+++ graphql_compiler/schema_generation/schema_graph.py
@@ -1,6 +1,7 @@
 # Copyright 2019-present Kensho Technologies, LLC.
 from abc import ABCMeta, abstractmethod
-from collections import namedtuple
+from collections import namedtuple, OrderedDict
+from itertools import chain
 
 import six
 
@@ -423,7 +424,7 @@ def _validate_property_names(class_name, properties):
                                           u'{}'.format(class_name, property_name))
 
 
-def get_subclass_sets_from_superclass_sets(superclass_sets):
+def _get_subclass_sets_from_superclass_sets(superclass_sets):
     """Return a dict mapping each class to its set of subclasses."""
     subclass_sets = dict()
     for subclass_name, superclass_names in six.iteritems(superclass_sets):
@@ -474,19 +475,99 @@ def link_schema_elements(elements, inheritance_structure):
             to_schema_element.in_connections.add(edge_class_name)
 
 
+
+class InheritanceStructure(object):
+    def __init__(self, class_to_immediate_superclasses):
+        toposorted_immediate_superclasses =_toposort_classes(class_to_immediate_superclasses)
+        self._superclass_sets = _get_transitive_superclass_sets(toposorted_immediate_superclasses)
+        self._subclass_sets = _get_subclass_sets_from_superclass_sets(self._superclass_sets)
+
+    @property
+    def superclass_sets(self):
+        return self._superclass_sets
+
+    @property
+    def subclass_sets(self):
+        return self._subclass_sets
+
+
+def _toposort_classes(name_to_superclasses):
+    """Return OrderedDict of class to superclasses where each class is before its subclasses."""
+    def get_class_topolist(class_name, processed_classes, current_trace):
+        """Return a topologically sorted list of this class's dependencies and class itself
+
+        Args:
+            class_name: string, name of the class to process
+            name_to_class: dict, class_name -> descriptor
+            current_trace: list of strings, list of classes traversed during the recursion
+
+        Returns:
+            list of dicts, list of class names sorted in topological order
+        """
+        # Check if this class has already been handled
+        if class_name in processed_classes:
+            return []
+
+        if class_name in current_trace:
+            raise AssertionError(
+                'Encountered self-reference in dependency chain of {}'.format(class_name))
+
+        class_list = []
+        # Recursively process superclasses
+        current_trace.add(class_name)
+        for superclass_name in name_to_superclasses[class_name]:
+            class_list.extend(get_class_topolist(superclass_name, processed_classes, current_trace))
+        current_trace.remove(class_name)
+        # Do the bookkeeping
+        class_list.append(class_name)
+        processed_classes.add(class_name)
+
+        return class_list
+
+    toposorted = []
+    for name in name_to_superclasses.keys():
+        toposorted.extend(get_class_topolist(name, set(), set()))
+    return OrderedDict((class_name, name_to_superclasses[class_name])
+                       for class_name in toposorted)
+
+
+def _get_transitive_superclass_sets(toposorted_class_to_immediate_superclasses):
+    """Return the transitive superclass sets from the toposorted class to superclass OrderedDict."""
+    # For each class name, construct its superclass set:
+    # itself + the set of class names from which it inherits.
+    superclass_sets = dict()
+    for class_name, immediate_superclass_names in six.iteritems(
+            toposorted_class_to_immediate_superclasses):
+        superclass_set = set(immediate_superclass_names)
+        superclass_set.add(class_name)
+
+        # Since the input data must be in topological order, the superclasses of
+        # the current class should have already been processed.
+        # A KeyError on the following line would mean that the input
+        # was not topologically sorted.
+        superclass_set.update(chain.from_iterable(
+            superclass_sets[superclass_name]
+            for superclass_name in immediate_superclass_names
+        ))
+
+        # Freeze the superclass set so it can't ever be modified again.
+        superclass_sets[class_name] = frozenset(superclass_set)
+    return superclass_sets
+
+
 # A way to describe a property's type and associated information:
 #   - type: GraphQLType, the type of this property
 #   - default: the default value for the property, used when a record is inserted without an
 #              explicit value for this property. Set to None if no default is given in the schema.
 PropertyDescriptor = namedtuple('PropertyDescriptor', ('type', 'default'))
 
-# A way to describe a schema's normalized inheritance structure.
-#   - superclass_sets: a dict, string -> set of strings, mapping each class to its
-#                      superclasses. The set of superclasses includes the class itself and
-#                      the transitive superclasses. For instance, if A is a superclass of B,
-#                      and B is a superclass of C, then C's superclass set is {'A', 'B', 'C'}.
-#   - subclass_sets: a dict, string -> set of strings, that is similarly defined.
-InheritanceStructure = namedtuple('InheritanceStructure', ('superclass_sets', 'subclass_sets'))
+# # A way to describe a schema's normalized inheritance structure.
+# #   - superclass_sets: a dict, string -> set of strings, mapping each class to its
+# #                      superclasses. The set of superclasses includes the class itself and
+# #                      the transitive superclasses. For instance, if A is a superclass of B,
+# #                      and B is a superclass of C, then C's superclass set is {'A', 'B', 'C'}.
+# #   - subclass_sets: a dict, string -> set of strings, that is similarly defined.
+# InheritanceStructure = namedtuple('InheritanceStructure', ('superclass_sets', 'subclass_sets'))
 
 # A way to describe an index:
 #   - name: string, the name of the index.
